requirements.txt
pika==1.3.2
redis==5.0.1
pymongo==4.6.1
boto3==1.34.34
pyyaml==6.0.1
numpy==1.24.3
